pyproject.toml
[tool.poetry]
name = "aegis-core"
version = "0.1.0"
description = "Aegis Core - Proactive API Security Gateway"
authors = ["Jules"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
httpx = "^0.25.1"
grpcio = "^1.59.2"
grpcio-tools = "^1.59.2"
protobuf = "^4.25.1"
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
cryptography = "^41.0.5"
redis = "^5.0.1"
python-json-logger = "^2.0.7"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
